requirements.txt
aiogram
python-dotenv
openai
apscheduler
httpx
yt_dlp
groq
gigachat
beautifulsoup4
pypdf
SpeechRecognition
pydub
pyrogram>=2.0.0
matplotlib
qrcode[pil]
opencv-python-headless
numpy
docx2pdf
mammoth
xhtml2pdf
